add format to currency helper
