+++ src/components/ShoppingCart/ShoppingCart.styled.js
@@ -1,0 +1,1 @@
+import styled from 'styled-components'
