+++ src/helpers/formatNumberToCurrency.js
@@ -1,0 +1,3 @@
+export default function formatCurrency(num) {
+  return '$' + num.toFixed(2).replace(/\d(?=(\d{3})+\.)/g, '$&,')
+}
