+++ src/pages/cart/index.js
@@ -1,22 +1,5 @@
-import { useShopStore } from '../../store'
+import ShoppingCart from '../../components/ShoppingCart'
 
 export default function CartPage() {
-  const cart = useShopStore((state) => state.cart)
-  /*  const currentProduct = cart[0].id */
-
-  return (
-    <>
-      <h1>Cart</h1>
-      <ul>
-        {cart && cart.map((item) => (
-          <li key={item.id}>
-            <p>{item.name}</p>
-            <p>{item.price}</p>
-            <p>{item.quantity}</p>
-            <p>{item.sum}</p>
-          </li>
-        ))}
-      </ul>
-    </>
-  )
+  return <ShoppingCart />
 }
